introduce shadow node
