+++ src/nodes/TSL.js
@@ -130,6 +130,7 @@ export * from './gpgpu/AtomicFunctionNode.js';
 export * from './accessors/Lights.js';
 export * from './lighting/LightsNode.js';
 export * from './lighting/LightingContextNode.js';
+export * from './lighting/ShadowNode.js';
 
 // pmrem
 export * from './pmrem/PMREMNode.js';
